requirements.txt
fastapi
uvicorn
yt-dlp
youtube-transcript-api
pywhispercpp
imageio-ffmpeg
python-dotenv
pydantic
google-generativeai
